WIP - Splitting temp_dir into LRD and CCM to resolve merging LRD files issues.

+++ utils/post_processing.py
@@ -3,15 +3,15 @@ import re
 import datetime
 import pikepdf
 
-# TODO: 6-26-23 - test refactored driver function w/ LRD and CCM
+def delete_pdf_files(output_directory, merged_file_path):
+    if not os.path.exists(merged_file_path):
+        print("Merged file does not exist. Not deleting any files.")
+        return False
 
-def delete_pdf_files(directory_path):
     files_deleted = False
-
-    for file_name in os.listdir(directory_path):
-        print(f'filename: {file_name}')
+    for file_name in os.listdir(output_directory):
         if file_name.endswith('.pdf'):
-            file_path = os.path.join(directory_path, file_name)
+            file_path = os.path.join(output_directory, file_name)
             os.remove(file_path)
             files_deleted = True
 
@@ -34,26 +34,20 @@ def extract_lrd_data(pdf_file):
     return None, None
 
 
-def extract_pdf_data(temp_dir):
+def extract_pdf_data(temp_dir, prefix):
     today = datetime.date.today().strftime('%m-%d-%y')
     pdf_files = os.listdir(temp_dir)
-    pdf_data_ccm = []
-    pdf_data_lrd = []
+    pdf_data = []
     for pdf_file in pdf_files:
         if pdf_file.endswith('.pdf'):
             file_path = os.path.join(temp_dir, pdf_file)
-            if pdf_file.startswith('CCM'):
-                regex_num, total_amount = extract_ccm_data(pdf_file)
-                if regex_num is not None and total_amount is not None:
-                    pdf_data_ccm.append((regex_num, today, total_amount, file_path))
-            elif pdf_file.startswith('LRD'):
-                regex_num, _ = extract_lrd_data(pdf_file)
+            if pdf_file.startswith(prefix):
+                regex_num, total_amount = extract_ccm_data(pdf_file) if prefix == 'CCM' else extract_lrd_data(pdf_file)
                 if regex_num is not None:
-                    pdf_data_lrd.append((regex_num, today, None, file_path))
-    pdf_data_ccm.sort(key=lambda x: x[0])
-    pdf_data_lrd.sort(key=lambda x: x[0])
-    total_amount_sum = round(sum(item[2] for item in pdf_data_ccm), 2)
-    return pdf_data_ccm, pdf_data_lrd, total_amount_sum
+                    pdf_data.append((regex_num, today, total_amount, file_path))
+    pdf_data.sort(key=lambda x: x[0])
+    total_amount_sum = round(sum(item[2] for item in pdf_data if item[2] is not None), 2) if prefix == 'CCM' else None
+    return pdf_data, total_amount_sum
 
 def check_file_exists(output_path):
     file_path = os.path.join(output_path)
@@ -69,36 +63,39 @@ def merge_pdfs(pdf_data):
 
 def save_merged_pdf(temp_dir, merged_pdf, total_amount_sum, file_prefix):
     today = datetime.date.today().strftime('%m-%d-%y')
-    new_file_name = ''
-    if file_prefix == 'CCM':
-        new_file_name = f'{file_prefix}-{today}-{total_amount_sum}.pdf'
-    elif file_prefix == 'LRD':
-        new_file_name = f'{today}-Loyalty.pdf'
-    output_dir = temp_dir[:-5]
-    output_path = os.path.join(output_dir, new_file_name)
+    new_file_name = f'{file_prefix}-{today}-{total_amount_sum}.pdf' if file_prefix == 'CCM' else f'{today}-Loyalty.pdf'
+    output_path = os.path.join(temp_dir, new_file_name)
     merged_pdf.save(output_path)
     merged_pdf.close()
-    merged_file_exists = check_file_exists(output_path)
-    if merged_file_exists:
-        if file_prefix == 'CCM':
-            print(f'EXXON CCM PDFs have been merged, renamed "{new_file_name}" and moved to: {output_path}\nDeleting temporary PDF files in {temp_dir}')
-        elif file_prefix == 'LRD':
-            print(f'Loyalty PDFs have been merged and saved as "{new_file_name}" to: {output_path}\nDeleting temporary PDF files in {temp_dir}')
-        temp_files_deleted = delete_pdf_files(temp_dir)
-        if temp_files_deleted:
-            print('Temporary PDF files have been deleted.')
-        else:
-            print('Temporary PDF files were not deleted.')
-    else:
-        print('Failed to save the merged PDF.')
 
-    merged_pdf.close()
+    while not os.path.exists(output_path):
+        print("Waiting for merged file to be created...")
+        time.sleep(1)
+
+    print(
+        f'{file_prefix} PDFs have been merged, renamed "{new_file_name}" and moved to: {output_path}\nDeleting temporary PDF files in {temp_dir}')
+    temp_files_deleted = delete_pdf_files(temp_dir, output_path)
+    if temp_files_deleted:
+        print('Temporary PDF files have been deleted.')
+    else:
+        print('Temporary PDF files were not deleted.')
+
 
 def merge_rename_and_summate(temp_dir):
-    pdf_data_ccm, pdf_data_lrd, total_amount_sum_ccm = extract_pdf_data(temp_dir)
+    ccm_temp_dir = os.path.join(temp_dir, "CCM")
+    lrd_temp_dir = os.path.join(temp_dir, "LRD")
 
+
+    # Move the files to their respective directories
+    for file in os.listdir(temp_dir):
+        if file.startswith("CCM"):
+            os.rename(os.path.join(temp_dir, file), os.path.join(ccm_temp_dir, file))
+        elif file.startswith("LRD"):
+            os.rename(os.path.join(temp_dir, file), os.path.join(lrd_temp_dir, file))
+
+    pdf_data_ccm, total_amount_sum_ccm = extract_pdf_data(ccm_temp_dir, 'CCM')
     merged_pdf_ccm = merge_pdfs(pdf_data_ccm)
-    save_merged_pdf(temp_dir, merged_pdf_ccm, total_amount_sum_ccm, 'CCM')
+    save_merged_pdf(ccm_temp_dir, merged_pdf_ccm, total_amount_sum_ccm, 'CCM')
 
-    merged_pdf_lrd = merge_pdfs(pdf_data_lrd)
-    save_merged_pdf(temp_dir, merged_pdf_lrd, None, 'LRD')
+    pdf_data_lrd, _ = extract_pdf_data(lrd_temp_dir, 'LRD')
+
